Blas.Primitive.Types

Description

Enumerations used by the Blas interface.

Documentation

data Order

Memory layout of a matrix.

Constructors

RowMajor
Row-major order.

ColMajor
Column-major order.

data Transpose

Operation applied to a matrix.

Constructors

NoTrans
No operation.

Trans
Transposition.

ConjTrans
Conjugation and transposition.

data Uplo

Specifies which triangular part of a matrix is used.

Constructors

Upper
Upper triangular part.

Lower
Lower triangular part.

data Diag

Whether a matrix is unit triangular.

Constructors

NonUnit
Not unit triangular.

Unit
Unit triangular.

data Side

The side that a matrix appears in the product.

Constructors

Left
On the left side.

Right
On the right side.